+++ src/realm/alloc_slab.cpp
@@ -1126,21 +1126,7 @@ inline bool randomly_false_in_debug(bool x)
   which extends beyond the last page of a file, the result is undefined, so we can't do that.
   We don't want to extend the file in increments as large as the chunk size.
 
-  The approach chosen on unixes is to only reserve a *virtual address range* large enough for
-  a chunk, and then gradually re-map it as the file grows. Having reserved it first, we're
-  guaranteed that the remapping operation will succeed. This allows us to ensure that any
-  existing mapping continues to stay valid even as new mappings are added. This is crucial
-  for sharing the mapping between multiple readers.
-
-  Unfortunately, there are no similar guarantees on Windows. On Windows you *can* reserve
-  address space for a full chunk, but you cannot remap only a portion of it. Only the full
-  section. Furthermore, you have to first release the reserved address space and then map
-  the file. This conceivably allows a different thread asking for virtual memory to grab
-  it between it is released and the file is mapped. Another complication is that on Windows
-  you cannot subdivide or coalesce a mapping range, because established mappings are identified
-  solely by their starting address, not as a range.
-
-  On Windows we choose to grow by creating a new larger memory mapping, which replaces the
+  As the file grows, we grow the mapping by creating a new larger one, which replaces the
   old one in the mapping table. However, we must keep the old mapping open, because older
   read transactions will continue to use it. Hence, the replaced mappings are accumulated
   and only cleaned out once we know that no transaction can refer to them anymore.
@@ -1151,18 +1137,6 @@ inline bool randomly_false_in_debug(bool x)
   The binding to the file is done by software. This allows us to "cheat" and allocate
   entire sections. With encryption, it doesn't matter if the mapped memory logically
   extends beyond the end of file, because it will not be accessed.
-
-  The following "branching out" is placed in "file.cpp":
-
-  Operation:   Encryption:           No encryption, Unixes:     No encryption, Windows:
-
-  reserve      allocate req          reserve chunk              no-op
-  extend       no-op                 remap part of reservation  allocate req (*)
-  mmap         allocate req          allocate req               allocate req
-
-  (*) If the mapping is already established for part of the requested area, then
-      extend() fail, and we must preserve the earlier mapping as described above,
-      then create a new one.
 
   Growing/Changing the mapping table.
 
@@ -1198,7 +1172,7 @@ void SlabAlloc::update_reader_view(size_t file_size)
     REALM_ASSERT_DEBUG(is_free_space_clean());
     bool requires_new_translation = false;
 
-    // Extend mapping by adding sections, or by extending sections
+    // Extend mapping by adding sections, potentially replacing older sections
     size_t old_baseline = m_baseline.load(std::memory_order_relaxed);
     auto old_slab_base = align_size_to_section_boundary(old_baseline);
     size_t old_num_sections = get_section_index(old_slab_base);
@@ -1212,20 +1186,13 @@ void SlabAlloc::update_reader_view(size_t file_size)
         if (file_size < old_slab_base) {
             size_t section_start_offset = get_section_base(old_num_sections - 1);
             size_t section_size = file_size - section_start_offset;
-            auto ok = m_mappings[mapping_index].extend(m_file, File::access_ReadOnly, section_size);
-            ok = randomly_false_in_debug(ok);
-            if (!ok) {
-                requires_new_translation = true;
-                size_t section_reservation = get_section_base(old_num_sections) - section_start_offset;
-                // save the old mapping/keep it open
-                OldMapping oldie(m_youngest_live_version, m_mappings[mapping_index]);
-                m_old_mappings.emplace_back(std::move(oldie));
-                m_mappings[mapping_index].reserve(m_file, File::access_ReadOnly, section_start_offset,
-                                                  section_reservation);
-                ok = m_mappings[mapping_index].extend(m_file, File::access_ReadOnly, section_size);
-                m_mapping_version++;
-            }
-            REALM_ASSERT(ok);
+            requires_new_translation = true;
+            // save the old mapping/keep it open
+            OldMapping oldie(m_youngest_live_version, m_mappings[mapping_index]);
+            m_old_mappings.emplace_back(std::move(oldie));
+            m_mappings[mapping_index] =
+                util::File::Map<char>(m_file, section_start_offset, File::access_ReadOnly, section_size);
+            m_mapping_version++;
         }
         else { // extension stretches over multiple sections:
 
@@ -1234,20 +1201,16 @@ void SlabAlloc::update_reader_view(size_t file_size)
             if (old_baseline < old_slab_base) {
                 size_t section_start_offset = get_section_base(old_num_sections - 1);
                 size_t section_size = old_slab_base - section_start_offset;
-                auto ok = m_mappings[mapping_index].extend(m_file, File::access_ReadOnly, section_size);
-                ok = randomly_false_in_debug(ok);
-                if (!ok) {
-                    // we could not extend the old mapping, so replace it with a full, new one
-                    requires_new_translation = true;
-                    size_t section_reservation = get_section_base(old_num_sections) - section_start_offset;
-                    REALM_ASSERT(section_size == section_reservation);
-                    // save the old mapping/keep it open
-                    OldMapping oldie(m_youngest_live_version, m_mappings[mapping_index]);
-                    m_old_mappings.emplace_back(std::move(oldie));
-                    m_mappings[mapping_index] =
-                        util::File::Map<char>(m_file, section_start_offset, File::access_ReadOnly, section_size);
-                    m_mapping_version++;
-                }
+                // we could not extend the old mapping, so replace it with a full, new one
+                requires_new_translation = true;
+                size_t section_reservation = get_section_base(old_num_sections) - section_start_offset;
+                REALM_ASSERT(section_size == section_reservation);
+                // save the old mapping/keep it open
+                OldMapping oldie(m_youngest_live_version, m_mappings[mapping_index]);
+                m_old_mappings.emplace_back(std::move(oldie));
+                m_mappings[mapping_index] =
+                    util::File::Map<char>(m_file, section_start_offset, File::access_ReadOnly, section_size);
+                m_mapping_version++;
             }
 
             // 2. add any full mappings
@@ -1279,14 +1242,9 @@ void SlabAlloc::update_reader_view(size_t file_size)
                 REALM_ASSERT(num_mappings == num_full_mappings + 1);
                 size_t section_start_offset =
                     get_section_base(num_full_mappings + m_sections_in_compatibility_mapping);
-                size_t section_reservation =
-                    get_section_base(num_full_mappings + 1 + m_sections_in_compatibility_mapping) -
-                    section_start_offset;
                 size_t section_size = file_size - section_start_offset;
                 util::File::Map<char> mapping;
-                mapping.reserve(m_file, File::access_ReadOnly, section_start_offset, section_reservation);
-                auto ok = mapping.extend(m_file, File::access_ReadOnly, section_size);
-                REALM_ASSERT(ok); // should allways succeed, as this is the first extend()
+                mapping = util::File::Map<char>(m_file, section_start_offset, File::access_ReadOnly, section_size);
                 m_mappings[num_full_mappings] = std::move(mapping);
             }
         }
@@ -1371,6 +1329,7 @@ void SlabAlloc::rebuild_translations(bool requires_new_translation, size_t old_n
         new_translation_table = new RefTranslation[m_translation_table_size];
         for (int i = 0; i < m_sections_in_compatibility_mapping; ++i) {
             new_translation_table[i].mapping_addr = m_compatibility_mapping.get_addr() + get_section_base(i);
+            REALM_ASSERT(new_translation_table[i].mapping_addr);
 #if REALM_ENABLE_ENCRYPTION
             new_translation_table[i].encrypted_mapping = m_compatibility_mapping.get_encrypted_mapping();
 #endif
@@ -1380,6 +1339,7 @@ void SlabAlloc::rebuild_translations(bool requires_new_translation, size_t old_n
     for (size_t k = old_num_sections; k < num_mappings; ++k) {
         auto i = k + m_sections_in_compatibility_mapping;
         new_translation_table[i].mapping_addr = m_mappings[k].get_addr();
+        REALM_ASSERT(new_translation_table[i].mapping_addr);
 #if REALM_ENABLE_ENCRYPTION
         new_translation_table[i].encrypted_mapping = m_mappings[k].get_encrypted_mapping();
 #endif
@@ -1387,6 +1347,7 @@ void SlabAlloc::rebuild_translations(bool requires_new_translation, size_t old_n
     for (size_t k = 0; k < free_space_size; ++k) {
         char* base = m_slabs[k].addr;
         auto i = num_mappings + m_sections_in_compatibility_mapping + k;
+        REALM_ASSERT(base);
 #if REALM_ENABLE_ENCRYPTION
         new_translation_table[i] = {base, nullptr};
 #else